backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
langchain
langchain-core
langchain-community
langchain-google-genai
chromadb
sentence-transformers
huggingface-hub